Reduce parameter fanout
